Split SharedPreferenceStorage and add internal modifier.
